requirements.txt
streamlit
seaborn
pandas
numpy
matplotlib
joblib
scikit-learn==1.0.2
plotly
Flask
tensorflow-cpu
gunicorn